Replace std::clock by std::chrono in canvas for better precision.
Former-commit-id: 4bc16ef1fb95efdb56e0903889f8ac348696ee48

+++ tutorial/02-handle-canvas-resizing/src/Main.cpp
@@ -60,7 +60,7 @@ main(int argc, char** argv)
             camera->component<PerspectiveCamera>()->aspectRatio((float) width / (float) height);
         });
 
-        auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint t, uint dt)
+        auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint t, float dt)
         {
             cube->component<Transform>()->matrix()->prependRotationY(.01f);
             sceneManager->nextFrame();
